fp32 final sum in TE MoE path (minor consistency)
Exhaustive verification shows every TE MoE step bitwise equivalent to its
analog (F.linear, bmm, scatter_add), but composition differs from ref by
L1=7.5e-4. Likely due to EP=4 sharding differences: ref's Megatron dispatches
experts across 4 EP ranks with inter-rank communication, while nano runs
all 144 experts on one GPU. The accumulation patterns of gate-weighted
expert outputs differ at bf16 ULP level.

Current per-layer L1 with all fixes + NANO_TE_MOE=1:
  block 0: 3.9e-5  ✓ (below target)
  block 1: 7.7e-4  (7x target, likely EP-dispatch artifact)
  block 8: 2.3e-3  (similar pattern)

End-to-end Test A Δ: +0.01373 (from +0.01387 baseline, -1% absolute)

+++ model.py
@@ -539,7 +539,8 @@ class MoEFFN(nn.Module):
                 out_perm, row_id_map, merging_probs=probs,
                 restore_shape=x_flat.shape, map_type='mask',
             )
-            out = (shared_out + routed_out).view(B, T, C)
+            # Fp32 sum then cast (matches nano's non-TE path)
+            out = (shared_out.float() + routed_out.float()).to(x_flat.dtype).view(B, T, C)
             # Return per Block.forward expectation: (out, aux)
             if self.training and self.seq_aux_alpha > 0:
                 # Skip aux for now when using TE path (only affects training)
